+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Add Warsaw (europe-central2) Cloud Function Location to Firebase Extension template.
+- Add Singapore (asia-southeast1) as a valid Firebase Realtime Database location.
 - Fix issue where `firebase init database` overwrites entire `firebase.json` (#3299)